src/cartadderagent/requirements.txt
a2a-sdk[http-server]
google-adk
grpcio
grpcio-tools
protobuf
requests
httpx
uvicorn
fastapi
